// src/app/sistema/Servicos/cursos/enums/modulos.ts
export enum Modulos {
    ONBOARDING = 'ONBOARDING',
    HISTORIA_SUCESSO = 'HISTORIA_SUCESSO',
    PRINCIPIOS_OTICA = 'PRINCIPIOS_OTICA',
    SCRIPT_VENDAS = 'SCRIPT_VENDAS',
    CONSEGUIR_CLIENTES = 'CONSEGUIR_CLIENTES',
    CONSULTOR_ALTA_PERFORMANCE = 'CONSULTOR_ALTA_PERFORMANCE',
    LIMPEZA_MANUTENCAO = 'LIMPEZA_MANUTENCAO',
    MANUTENCAO_OCULOS = 'MANUTENCAO_OCULOS',
    GARANTIA_PRODUTOS = 'GARANTIA_PRODUTOS',
    EMBALAGEM_PADRAO = 'EMBALAGEM_PADRAO',
    ENTREGA_OCULOS_GRAU = 'ENTREGA_OCULOS_GRAU',
    PADROES_ATENDIMENTO = 'PADROES_ATENDIMENTO',
    SSOTICA_SISTEMA_VENDAS = 'SSOTICA_SISTEMA_VENDAS',
    SSOTICA_CAIXA = 'SSOTICA_CAIXA',
    EU_SOU_VENDEDOR = 'EU_SOU_VENDEDOR',
    INTELIGENCIA_EMOCIONAL = 'INTELIGENCIA_EMOCIONAL',
}